# fix(perf): lazy load un-common dependencies for npm run