--- a/lib/normalize.js
+++ b/lib/normalize.js
@@ -1,11 +1,8 @@
 const semver = require('semver')
 const fs = require('fs/promises')
 const { glob } = require('glob')
-const legacyFixer = require('normalize-package-data/lib/fixer.js')
-const legacyMakeWarning = require('normalize-package-data/lib/make_warning.js')
 const path = require('path')
 const log = require('proc-log')
-const git = require('@npmcli/git')
 const hostedGitInfo = require('hosted-git-info')
 
 // used to be npm-normalize-package-bin
@@ -333,6 +330,7 @@ const normalize = async (pkg, { strict, steps, root, changes, allowLegacyCase })
 
   // populate "gitHead" attribute
   if (steps.includes('gitHead') && !data.gitHead) {
+    const git = require('@npmcli/git')
     const gitRoot = await git.find({ cwd: pkg.path, root })
     let head
     if (gitRoot) {
@@ -518,6 +516,8 @@ const normalize = async (pkg, { strict, steps, root, changes, allowLegacyCase })
   }
 
   if (steps.includes('normalizeData')) {
+    const legacyFixer = require('normalize-package-data/lib/fixer.js')
+    const legacyMakeWarning = require('normalize-package-data/lib/make_warning.js')
     legacyFixer.warn = function () {
       changes?.push(legacyMakeWarning.apply(null, arguments))
     }