I know we avoid lazy loading but I think this case is worthy:

Before:

```bash
$ hyperfine --warmup 3 "node ./bin/npm-cli.js run empty"
Benchmark 1: node ./bin/npm-cli.js run empty
  Time (mean ± σ):     197.3 ms ±   3.5 ms    [User: 183.1 ms, System: 51.9 ms]
  Range (min … max):   193.2 ms … 207.5 ms    15 runs
```

After:

```bash
$ hyperfine --warmup 3 "node ./bin/npm-cli.js run empty"
Benchmark 1: node ./bin/npm-cli.js run empty
  Time (mean ± σ):     184.6 ms ±   2.3 ms    [User: 170.1 ms, System: 51.1 ms]
  Range (min … max):   181.6 ms … 190.5 ms    16 runs
```