@wraithgar 2024-04-06
We need to be 100% sure this code path isn't hit when running `npm i -g npm`