@wraithgar 2024-04-07
Additionally, thinking of potential future problems.  We are moving in the direction of doing *fewer* normalizations, not more.  So the likelihood of us adding either of those steps during install is very low.