@wraithgar 2024-04-07
Eventually `normalize-package-data` needs to be brought under the umbrella of this package.  Eventually.